+++ src/main.rs
@@ -1,36 +1,17 @@
 #[macro_use]
 extern crate pest_derive;
-#[macro_use]
-extern crate lazy_static;
 extern crate pest;
-use map::{DefinitionMap, Types};
 use parser::ast;
 use rustyline::error::ReadlineError;
 use rustyline::Editor;
-use std::sync::Mutex;
 
-mod map;
 mod parser;
-
-lazy_static! {
-    static ref DEFINITIONS: Mutex<DefinitionMap> = {
-        let mut definition_map = DefinitionMap::new();
-
-        definition_map.insert(
-            "version".to_string(),
-            Types::String(env!("CARGO_PKG_VERSION").to_string()),
-        );
-        definition_map.insert(
-            "author".to_string(),
-            Types::String(env!("CARGO_PKG_AUTHORS").to_string()),
-        );
-        definition_map.insert("license".to_string(), Types::String("GPL-3.0".to_string()));
-        Mutex::new(definition_map)
-    };
-}
 
 fn main() {
     let mut rl = Editor::<()>::new();
+
+    rl.load_history("history.txt")
+        .unwrap_or_else(|_| println!("Coudn't run history"));
 
     loop {
         let readline = rl.readline("🎆 >> ");
@@ -38,7 +19,7 @@ fn main() {
             Ok(line) => {
                 if !line.is_empty() {
                     rl.add_history_entry(&line);
-                    println!("{:?}", ast::parse(&line));
+                    println!("{:#?}", ast::parse(&line));
                 }
             }
             Err(ReadlineError::Interrupted) => {